requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
psycopg2-binary==2.8.4
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.0.1
drf-yasg==1.17.1